+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- ✨ add pangram support to word filtering(pr [#68])
+
 ## [0.2.0] - 2025-04-08
 
 ### Added
@@ -234,6 +240,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#65]: https://github.com/jerus-org/wpsr/pull/65
 [#66]: https://github.com/jerus-org/wpsr/pull/66
 [#67]: https://github.com/jerus-org/wpsr/pull/67
+[#68]: https://github.com/jerus-org/wpsr/pull/68
+[Unreleased]: https://github.com/jerus-org/wpsr/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/jerus-org/wpsr/compare/v0.1.14...v0.2.0
 [0.1.14]: https://github.com/jerus-org/wpsr/compare/v0.1.13...v0.1.14
 [0.1.13]: https://github.com/jerus-org/wpsr/compare/v0.1.12...v0.1.13